ProyectoClasificode/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
psycopg>=3.1.12
pandas==2.1.3
PyJWT==2.8.0
alembic==1.12.1
python-dotenv==1.0.0
pgvector==0.2.4
passlib[bcrypt]==1.7.4
numpy==1.26.4
rapidfuzz==3.4.0
openai==1.3.0
sentence-transformers==2.2.2
spacy==3.7.2
requests>=2.31.0
selenium>=4.14.0
pdfminer.six>=20221105